Visual Studio doesn't like two CPP files with the same name
