+++ modules/c++/sio.lite/source/SioFileReader.cpp
