+++ modules/c++/sio.lite/source/SioFileWriter.cpp
