+++ modules/c++/tiff/source/TiffFileWriter.cpp
