+++ modules/c++/tiff/wscript
@@ -1,5 +1,4 @@
 NAME            = 'tiff'
-MAINTAINER      = 'jmrandol@users.sourceforge.net'
 VERSION         = '1.0'
 MODULE_DEPS     = 'mt io'
